[PP][OPS-10176] more code cleanup
