+++ app/flow/FallBackFlow.scala
@@ -14,26 +14,23 @@
  * limitations under the License.
  */
 
-package behaviourspc
+package flow
 
-import models.InteractionCategories.CardReader
+import models.InteractionCategories.{CardReader, OnlineCategory}
 import models.TranResults.SuccessResult
 import models._
 
-class BinCheckCardDiscardedFlow(spcFlow: FlowData) extends Flow {
+class FallBackFlow(spcFlow: FlowData) extends Flow {
 
   val initialBehaviour: SpcBehaviour = handlePedLogOn
 
-  private lazy val  handlePedLogOn: SpcBehaviour = behave {
+  private lazy val handlePedLogOn: SpcBehaviour = behave {
     case pedLogOn: PedLogOn =>
       val pedLogOnResponse: SpcResponseMessage = PedLogOnResponse(HeaderNode(), pedLogOn.messageNode, SuccessResult, ErrorsNode(Seq.empty))
-      (
-        List(pedLogOnResponse),
-        handleSubmitPayment orElse CommonBehaviours.handlePedLogOff
-      )
+      (List(pedLogOnResponse), handleSubmitPayment orElse CommonBehaviours.handlePedLogOff)
   }
 
-  private lazy val  handleSubmitPayment: SpcBehaviour = behave{
+  private lazy val handleSubmitPayment: SpcBehaviour = behave {
     case submitPayment: SubmitPayment =>
       val paymentSubmittedData = SubmittedData(
         totalAmount         = submitPayment.transactionNode.amountNode.totalAmount,
@@ -44,17 +41,15 @@ class BinCheckCardDiscardedFlow(spcFlow: FlowData) extends Flow {
       )
 
       val submitPaymentResponse = SubmitPaymentResponse(HeaderNode(), submitPayment.messageNode, SuccessResult)
-      (
-        List(submitPaymentResponse),
-        handleProcessTransaction(paymentSubmittedData)
-      )
+      (List(submitPaymentResponse), handleProcessTransaction(paymentSubmittedData))
   }
 
   //sends UpdatePaymentEnhanced
   private def handleProcessTransaction(submittedData: SubmittedData): SpcBehaviour = behave {
     case processTransaction: ProcessTransaction =>
+
       //Display sequence - card validation
-      val interimResponses: Seq[PosDisplayMessage] = spcFlow.displayMessagesValidation.map{
+      val interimResponses = spcFlow.displayMessagesValidation.map{
         case (interactionEvents, interactionPrompt) =>
           val interactionNode = InteractionNode(category = CardReader, event = interactionEvents, prompt = interactionPrompt)
           val posDisplayMessageInsertCard = PosDisplayMessage(HeaderNode(), processTransaction.messageNode, interactionNode, SuccessResult, ErrorsNode(Seq.empty))
@@ -66,15 +61,40 @@ class BinCheckCardDiscardedFlow(spcFlow: FlowData) extends Flow {
       val transactionNode = TransactionNode(amountNode = amountNode)
       val cardNode = UpeCardNode(spcFlow.paymentCard)
       val updatePaymentEnhanced = UpdatePaymentEnhanced(HeaderNode(), processTransaction.messageNode, transactionNode, cardNode, SuccessResult, ErrorsNode(Seq.empty))
-
       (
         interimResponses :+[SpcResponseMessage] updatePaymentEnhanced,
-        handleTransactionCancelled(submittedData)
+        handleUpdatePaymentEnhancedResponse(submittedData) orElse handleTransactionCancelled(submittedData) orElse handleGetTransactionDetails(submittedData)
+      )
+  }
+
+  private def handleUpdatePaymentEnhancedResponse(submittedData: SubmittedData): SpcBehaviour = behave {
+    case updatePaymentEnhancedResponse: UpdatePaymentEnhancedResponse =>
+      val finalAmount = updatePaymentEnhancedResponse.amountNode.finalAmountO
+      val totalAmount = updatePaymentEnhancedResponse.amountNode.totalAmount
+
+      //Display sequence - card Authentication
+      val interimResponses = spcFlow.displayMessagesAuthentication.map {
+        case (interactionEvents, interactionPrompt) =>
+          val interactionNode = InteractionNode(category = OnlineCategory, event = interactionEvents, prompt = interactionPrompt)
+          val posDisplayMessageInsertCard = PosDisplayMessage(HeaderNode(), updatePaymentEnhancedResponse.messageNode, interactionNode, SuccessResult, ErrorsNode(Seq.empty))
+          posDisplayMessageInsertCard
+      }
+
+      //PosPrintReceipt client
+      val merchantReceiptNode = ReceiptMerchantNode(spcFlow, submittedData, totalAmount, finalAmount)
+      val posPrintReceipt = PosPrintReceipt(HeaderNode(), updatePaymentEnhancedResponse.messageNode, merchantReceiptNode, SuccessResult, ErrorsNode(Seq.empty))
+
+      val posDecisionTransNode = PdTransNode(TransactionDecisions.SignatureRequired)
+      val posDecisionMessage = PosDecisionMessage(HeaderNode(), updatePaymentEnhancedResponse.messageNode, posDecisionTransNode)
+      (
+        interimResponses :+[SpcResponseMessage] posPrintReceipt :+ posDecisionMessage,
+        handleTransactionCancelled(submittedData) orElse handleGetTransactionDetails(submittedData)
       )
   }
 
   private def handleTransactionCancelled(submittedData: SubmittedData): SpcBehaviour = behave {
     case cancelTransaction: CancelTransaction =>
+
       //processTransactionResponse
       val amountNode = AmountNode(submittedData.totalAmount, submittedData.currency, submittedData.country, None)
 
@@ -88,6 +108,31 @@ class BinCheckCardDiscardedFlow(spcFlow: FlowData) extends Flow {
       val processTransactionResponse = ProcessTransactionResponse(
         headerNode           = HeaderNode(),
         messageNode          = cancelTransaction.messageNode,
+        ptrTransactionNode   = ptrTransactionNode,
+        ptrCardNode          = cardNode,
+        result               = spcFlow.transactionResult,
+        paymentResult        = PaymentResults.cancelled,
+        receiptNodeCustomerO = None,
+        receiptNodeMerchantO = None,
+        errorsNode           = ErrorsNode(Seq.empty))
+      (List(processTransactionResponse), CommonBehaviours.handleFinalise)
+  }
+
+  private def handleGetTransactionDetails(submittedData: SubmittedData): SpcBehaviour = behave {
+    case getTransactionDetails: GetTransactionDetails =>
+      //processTransactionResponse
+      val amountNode = AmountNode(submittedData.totalAmount, submittedData.currency, submittedData.country, None)
+
+      val ptrTransactionNode = PtrTransactionNode(
+        amountNode      = amountNode,
+        verification    = spcFlow.cardVerificationMethod,
+        transactionDate = StubUtil.formatTransactionDate(submittedData.transactionDateTime),
+        transactionTime = StubUtil.formatTransactionTime(submittedData.transactionDateTime))
+      val cardNode = PtrResponseCardNode(spcFlow.paymentCard)
+
+      val processTransactionResponse = GetTransactionDetailsResponse(
+        headerNode           = HeaderNode(),
+        messageNode          = getTransactionDetails.messageNode,
         ptrTransactionNode   = ptrTransactionNode,
         ptrCardNode          = cardNode,
         result               = spcFlow.transactionResult,
