+++ app/flow/FlowData.scala
@@ -14,14 +14,18 @@
  * limitations under the License.
  */
 
-import behaviour.{BDefined, BDone, Behaviour}
-import models.{SpcRequestMessage, SpcResponseMessage}
+package flow
 
-package object behaviourspc {
+import models._
 
-  type SpcBehaviour = Behaviour[SpcRequestMessage, Seq[SpcResponseMessage]]
-
-  def behave(pf: PartialFunction[SpcRequestMessage, (Seq[SpcResponseMessage], SpcBehaviour)]): SpcBehaviour = BDefined(pf)
-  val done: Behaviour[SpcRequestMessage, Nothing] = BDone
-}
+final case class FlowData(
+    paymentCard:                   PaymentCard,
+    paymentResult:                 PaymentResult,
+    receiptNodeName:               ReceiptTypeName,
+    transactionResult:             TranResult,
+    cardVerificationMethod:        CardVerificationMethod,
+    transactionSource:             TransactionSource,
+    displayMessagesValidation:     Seq[(InteractionEvent, InteractionPrompt)],
+    displayMessagesAuthentication: Seq[(InteractionEvent, InteractionPrompt)]
+)
