+++ app/flow/NoReceiptFlow.scala
@@ -14,23 +14,27 @@
  * limitations under the License.
  */
 
-package behaviourspc
+package flow
 
-import models.InteractionCategories.CardReader
+import models.InteractionCategories.{CardReader, OnlineCategory}
 import models.TranResults.SuccessResult
 import models._
 
-class NoSurchargeFlow(spcFlow: FlowData, errorsNode: ErrorsNode) extends Flow {
+class NoReceiptFlow(flowData: FlowDataNoReceipt) extends Flow {
 
   val initialBehaviour: SpcBehaviour = handlePedLogOn
 
   private lazy val handlePedLogOn: SpcBehaviour = behave {
     case pedLogOn: PedLogOn =>
+
       val pedLogOnResponse: SpcResponseMessage = PedLogOnResponse(HeaderNode(), pedLogOn.messageNode, SuccessResult, ErrorsNode(Seq.empty))
-      (List(pedLogOnResponse), handleSubmitPayment orElse CommonBehaviours.handlePedLogOff)
+      (
+        List(pedLogOnResponse),
+        handleSubmitPayment orElse CommonBehaviours.handlePedLogOff
+      )
   }
 
-  private lazy val  handleSubmitPayment: SpcBehaviour = behave {
+  private def handleSubmitPayment: SpcBehaviour = behave {
     case submitPayment: SubmitPayment =>
       val paymentSubmittedData = SubmittedData(
         totalAmount         = submitPayment.transactionNode.amountNode.totalAmount,
@@ -44,91 +48,86 @@ class NoSurchargeFlow(spcFlow: FlowData, errorsNode: ErrorsNode) extends Flow {
       (List(submitPaymentResponse), handleProcessTransaction(paymentSubmittedData))
   }
 
-  //Do not send updatePaymentEnhanced but got to print message instead
+  //sends UpdatePaymentEnhanced
   private def handleProcessTransaction(submittedData: SubmittedData): SpcBehaviour = behave {
-
     case processTransaction: ProcessTransaction =>
-
       //Display sequence - card validation
-      val interimMessages = spcFlow.displayMessagesValidation.map{
+      val interimMessages = flowData.displayMessagesValidation.map{
         case (interactionEvents, interactionPrompt) =>
           val interactionNode = InteractionNode(category = CardReader, event = interactionEvents, prompt = interactionPrompt)
           val posDisplayMessageInsertCard = PosDisplayMessage(HeaderNode(), processTransaction.messageNode, interactionNode, SuccessResult, ErrorsNode(Seq.empty))
           posDisplayMessageInsertCard
       }
 
-      //PosPrintReceipt client
-      val merchantReceiptNode = ReceiptMerchantNode(spcFlow, submittedData, submittedData.totalAmount, None)
-      val posPrintReceipt: PosPrintReceipt = PosPrintReceipt(HeaderNode(), processTransaction.messageNode, merchantReceiptNode, SuccessResult, ErrorsNode(Seq.empty))
-      (
-        interimMessages :+[SpcResponseMessage] posPrintReceipt,
-        handlePosPrintReceiptResponse(submittedData, None, merchantReceiptNode) orElse handleTransactionCancelled(submittedData)
-      )
+      //UpdatePaymentEnhanced
+      val amountNode = AmountNode(submittedData.totalAmount, submittedData.currency, submittedData.country, None)
+      val transactionNode = TransactionNode(amountNode = amountNode)
+      val cardNode = UpeCardNode(flowData.paymentCard)
+      val updatePaymentEnhanced = UpdatePaymentEnhanced(HeaderNode(), processTransaction.messageNode, transactionNode, cardNode, SuccessResult, ErrorsNode(Seq.empty))
+      (interimMessages :+[SpcResponseMessage] updatePaymentEnhanced, handleUpdatePaymentEnhancedResponse(submittedData) orElse handleTransactionCancelled(submittedData))
   }
 
-  private def handlePosPrintReceiptResponse(submittedData: SubmittedData, finalAmount: Option[AmountInPence], merchantReceiptNode: ReceiptNode): SpcBehaviour = behave {
-    case posPrintReceiptResponse: PosPrintReceiptResponse =>
+  private def handleUpdatePaymentEnhancedResponse(submittedData: SubmittedData): SpcBehaviour = behave {
+    case updatePaymentEnhancedResponse: UpdatePaymentEnhancedResponse =>
+      val finalAmount = updatePaymentEnhancedResponse.amountNode.finalAmountO
+      val totalAmount = updatePaymentEnhancedResponse.amountNode.totalAmount
 
-      //PosPrintReceipt client
-      val clientReceiptNode = ReceiptNode.createReceiptNode(submittedData, spcFlow, submittedData.totalAmount, finalAmount)
+      //Display sequence - card Authentication
+      val interimMessages = flowData.displayMessagesAuthentication.map{
+        case (interactionEvents, interactionPrompt) =>
+          val interactionNode = InteractionNode(category = OnlineCategory, event = interactionEvents, prompt = interactionPrompt)
+          val posDisplayMessageInsertCard = PosDisplayMessage(HeaderNode(), updatePaymentEnhancedResponse.messageNode, interactionNode, SuccessResult, ErrorsNode(Seq.empty))
+          posDisplayMessageInsertCard
+      }
 
-      val posPrintReceipt = PosPrintReceipt(HeaderNode(), posPrintReceiptResponse.messageNode, clientReceiptNode, SuccessResult, ErrorsNode(Seq.empty))
-      (
-        List(posPrintReceipt),
-        handlePosPrintReceiptResponseWithPtr(submittedData, finalAmount, merchantReceiptNode, clientReceiptNode)
-      )
-
-  }
-
-  private def handlePosPrintReceiptResponseWithPtr(submittedData: SubmittedData, finalAmount: Option[AmountInPence], merchantReceiptNode: ReceiptNode, clientReceiptNode: ReceiptNode): SpcBehaviour = behave {
-    case posPrintReceiptResponse: PosPrintReceiptResponse =>
       //processTransactionResponse
-      val amountNode = AmountNode(submittedData.totalAmount, submittedData.currency, submittedData.country, finalAmount)
+      val amountNode = AmountNode(totalAmount, submittedData.currency, submittedData.country, finalAmount)
 
       val ptrTransactionNode = PtrTransactionNode(
         amountNode      = amountNode,
-        verification    = spcFlow.cardVerificationMethod,
+        verification    = flowData.cardVerificationMethod,
         transactionDate = StubUtil.formatTransactionDate(submittedData.transactionDateTime),
         transactionTime = StubUtil.formatTransactionTime(submittedData.transactionDateTime))
-      val cardNode = PtrResponseCardNode(spcFlow.paymentCard)
+      val cardNode = PtrResponseCardNode(flowData.paymentCard)
 
       val processTransactionResponse = ProcessTransactionResponse(
         headerNode           = HeaderNode(),
-        messageNode          = posPrintReceiptResponse.messageNode,
+        messageNode          = updatePaymentEnhancedResponse.messageNode,
         ptrTransactionNode   = ptrTransactionNode,
         ptrCardNode          = cardNode,
-        result               = spcFlow.transactionResult,
-        paymentResult        = spcFlow.paymentResult,
-        receiptNodeCustomerO = Some(clientReceiptNode),
-        receiptNodeMerchantO = Some(merchantReceiptNode),
-        errorsNode           = errorsNode)
-      (List(processTransactionResponse), CommonBehaviours.handleFinalise)
+        result               = flowData.transactionResult,
+        paymentResult        = flowData.paymentResult,
+        receiptNodeCustomerO = None,
+        receiptNodeMerchantO = None,
+        errorsNode           = ErrorsNode(Seq.empty))
+      (interimMessages :+[SpcResponseMessage] processTransactionResponse, CommonBehaviours.handleFinalise)
   }
 
   private def handleTransactionCancelled(submittedData: SubmittedData): SpcBehaviour = behave {
     case cancelTransaction: CancelTransaction =>
+
       //processTransactionResponse
       val amountNode = AmountNode(submittedData.totalAmount, submittedData.currency, submittedData.country, None)
 
       val ptrTransactionNode = PtrTransactionNode(
         amountNode      = amountNode,
-        verification    = spcFlow.cardVerificationMethod,
+        verification    = flowData.cardVerificationMethod,
         transactionDate = StubUtil.formatTransactionDate(submittedData.transactionDateTime),
         transactionTime = StubUtil.formatTransactionTime(submittedData.transactionDateTime))
-      val cardNode = PtrResponseCardNode(spcFlow.paymentCard)
+      val cardNode = PtrResponseCardNode(flowData.paymentCard)
 
       val processTransactionResponse = ProcessTransactionResponse(
         headerNode           = HeaderNode(),
         messageNode          = cancelTransaction.messageNode,
         ptrTransactionNode   = ptrTransactionNode,
         ptrCardNode          = cardNode,
-        result               = spcFlow.transactionResult,
+        result               = flowData.transactionResult,
         paymentResult        = PaymentResults.cancelled,
         receiptNodeCustomerO = None,
         receiptNodeMerchantO = None,
         errorsNode           = ErrorsNode(Seq.empty))
+
       (List(processTransactionResponse), CommonBehaviours.handleFinalise)
   }
-
 
 }
